+++ crates/stackable-operator/CHANGELOG.md
@@ -6,6 +6,14 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Extend `ObjectMetaBuilder` with `finalizers` ([#1094]).
+
+[#1094]: https://github.com/stackabletech/operator-rs/pull/1094
+
+## [0.97.0] - 2025-09-09
+
+### Added
+
 - BREAKING: Add a new CLI flag/env to disabling CRD maintenance: `--disable-crd-maintenance` ([#1085]).
 
 ### Changed